feat(ui): extract launcher rendering helpers into ui/launcher module
Move ~15 ImGui render helper functions from LauncherEditorShell.cpp's
anonymous namespace into a dedicated ui/launcher/LauncherWidgets module.

- Add ui/launcher/LauncherWidgets.h/.cpp with Horo::Launcher::UI namespace
- Keep icon-draw helpers (DrawTemplateTileIcon, DrawLauncherActionIcon,
  DrawSidebarFooterIcon) as static/private inside the .cpp anonymous namespace
- Expose DrawRecentProjectMetaIcon and ResolveLauncherVisualAsset publicly
  since they are called from class methods in LauncherEditorShell.cpp
- Add CenteredEmptyState to ui/common/HoroWidgets (shared widget)
- Update CMakeLists.txt to glob ui/launcher/*.cpp into ENGINE_SOURCES
- Replace all 23 inline call sites in LauncherEditorShell.cpp with
  Horo::Launcher::UI:: qualified names

LauncherEditorShell.cpp shrinks by ~500 lines; anonymous namespace now
contains only non-UI helpers (buffer utils, math, path predicates).
